+++ service/controller/resource/monitoring/remotewriteapiendpointconfigsecret/resource.go
@@ -3,28 +3,52 @@ package remotewriteapiendpointconfigsecret
 import (
 	"context"
 	"fmt"
-	"reflect"
 
 	"github.com/giantswarm/k8sclient/v7/pkg/k8sclient"
 	"github.com/giantswarm/microerror"
 	"github.com/giantswarm/micrologger"
-	pov1 "github.com/prometheus-operator/prometheus-operator/pkg/apis/monitoring/v1"
-	"gopkg.in/yaml.v2"
+	promv1 "github.com/prometheus-operator/prometheus-operator/pkg/apis/monitoring/v1"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"sigs.k8s.io/yaml"
 
+	"github.com/giantswarm/prometheus-meta-operator/v2/pkg/password"
 	"github.com/giantswarm/prometheus-meta-operator/v2/service/controller/resource/generic"
 	"github.com/giantswarm/prometheus-meta-operator/v2/service/key"
 )
 
 const (
-	Name = "remotewriteapiendpointconfigsecret"
+	Name                           = "remotewriteapiendpointconfigsecret"
+	remoteWriteEndpointTemplateURL = "https://%s/%s/api/v1/write"
 )
 
 type Config struct {
-	K8sClient  k8sclient.Interface
-	Logger     micrologger.Logger
-	BaseDomain string
+	K8sClient       k8sclient.Interface
+	Logger          micrologger.Logger
+	PasswordManager password.Manager
+	BaseDomain      string
+	Customer        string
+	Installation    string
+	Pipeline        string
+	Provider        string
+	Region          string
+}
+
+type RemoteWrite struct {
+	Name        string             `json:"name"`
+	Password    string             `json:"password"`
+	Username    string             `json:"username"`
+	URL         string             `json:"url"`
+	QueueConfig promv1.QueueConfig `json:"queueConfig"`
+}
+
+type GlobalRemoteWriteValues struct {
+	Global RemoteWriteValues `json:"global"`
+}
+
+type RemoteWriteValues struct {
+	RemoteWrite    []RemoteWrite     `json:"remoteWrite"`
+	ExternalLabels map[string]string `json:"externalLabels"`
 }
 
 func New(config Config) (*generic.Resource, error) {
@@ -34,10 +58,12 @@ func New(config Config) (*generic.Resource, error) {
 	}
 
 	c := generic.Config{
-		ClientFunc:    clientFunc,
-		Logger:        config.Logger,
-		Name:          Name,
-		GetObjectMeta: getObjectMeta,
+		ClientFunc: clientFunc,
+		Logger:     config.Logger,
+		Name:       Name,
+		GetObjectMeta: func(ctx context.Context, v interface{}) (metav1.ObjectMeta, error) {
+			return getObjectMeta(ctx, v, config.Provider)
+		},
 		GetDesiredObject: func(ctx context.Context, v interface{}) (metav1.Object, error) {
 			return toSecret(ctx, v, config)
 		},
@@ -51,21 +77,23 @@ func New(config Config) (*generic.Resource, error) {
 	return r, nil
 }
 
-func getObjectMeta(ctx context.Context, v interface{}) (metav1.ObjectMeta, error) {
+func getObjectMeta(ctx context.Context, v interface{}, provider string) (metav1.ObjectMeta, error) {
 	cluster, err := key.ToCluster(v)
 	if err != nil {
 		return metav1.ObjectMeta{}, microerror.Mask(err)
 	}
 
+	name, namespace := key.RemoteWriteAPIEndpointConfigSecretNameAndNamespace(cluster, provider)
+
 	return metav1.ObjectMeta{
-		Name:      key.RemoteWriteAgentConfigSecretName,
-		Namespace: key.Namespace(cluster),
+		Name:      name,
+		Namespace: namespace,
 		Labels:    key.PrometheusLabels(cluster),
 	}, nil
 }
 
 func toSecret(ctx context.Context, v interface{}, config Config) (*corev1.Secret, error) {
-	objectMeta, err := getObjectMeta(ctx, v)
+	objectMeta, err := getObjectMeta(ctx, v, config.Provider)
 	if err != nil {
 		return nil, microerror.Mask(err)
 	}
@@ -75,44 +103,65 @@ func toSecret(ctx context.Context, v interface{}, config Config) (*corev1.Secret
 		return nil, microerror.Mask(err)
 	}
 
-	remoteWrites := []pov1.RemoteWriteSpec{
+	config.Logger.Debugf(ctx, "generating password for the prometheus agent")
+	password, err := config.PasswordManager.GeneratePassword(32)
+	if err != nil {
+		config.Logger.Errorf(ctx, err, "failed to generate the prometheus agent password")
+		return nil, microerror.Mask(err)
+	}
+
+	config.Logger.Debugf(ctx, "generate password for the prometheus agent")
+
+	remoteWrites := []RemoteWrite{
 		{
-			URL: fmt.Sprintf("https://%s/%s/api/v1/write", config.BaseDomain, key.ClusterID(cluster)),
-			BasicAuth: &pov1.BasicAuth{
-				Username: corev1.SecretKeySelector{
-					LocalObjectReference: corev1.LocalObjectReference{
-						Name: key.RemoteWriteAgentSecretName,
-					},
-					Key: "username",
-				},
-				Password: corev1.SecretKeySelector{
-					LocalObjectReference: corev1.LocalObjectReference{
-						Name: key.RemoteWriteAgentSecretName,
-					},
-					Key: "password",
-				},
-			},
+			Name:        key.PrometheusMetaOperatorRemoteWriteName,
+			URL:         fmt.Sprintf(remoteWriteEndpointTemplateURL, config.BaseDomain, key.ClusterID(cluster)),
+			Username:    key.ClusterID(cluster),
+			Password:    password,
+			QueueConfig: defaultQueueConfig(),
 		},
 	}
 
-	marshalledRemoteWrites, err := yaml.Marshal(remoteWrites)
+	externalLabels := map[string]string{
+		key.ClusterIDKey:    key.ClusterID(cluster),
+		key.ClusterTypeKey:  key.ClusterType(config.Installation, cluster),
+		key.CustomerKey:     config.Customer,
+		key.InstallationKey: config.Installation,
+		key.PipelineKey:     config.Pipeline,
+		key.ProviderKey:     config.Provider,
+		key.RegionKey:       config.Region,
+	}
+
+	values := RemoteWriteValues{
+		RemoteWrite:    remoteWrites,
+		ExternalLabels: externalLabels,
+	}
+	marshalledValues, err := yaml.Marshal(GlobalRemoteWriteValues{values})
+
 	if err != nil {
 		return nil, microerror.Mask(err)
 	}
 
+	var immutable bool = true
 	secret := &corev1.Secret{
 		ObjectMeta: objectMeta,
 		Data: map[string][]byte{
-			"values": []byte(marshalledRemoteWrites),
+			"values": []byte(marshalledValues),
 		},
-		Type: "Opaque",
+		Type:      "Opaque",
+		Immutable: &immutable,
 	}
 	return secret, nil
 }
 
-func hasChanged(current, desired metav1.Object) bool {
-	c := current.(*corev1.Secret)
-	d := desired.(*corev1.Secret)
+func defaultQueueConfig() promv1.QueueConfig {
+	return promv1.QueueConfig{
+		Capacity:          30000,
+		MaxSamplesPerSend: 10000,
+		MaxShards:         10,
+	}
+}
 
-	return !reflect.DeepEqual(c.Data, d.Data)
+func hasChanged(current, desired metav1.Object) bool {
+	return false
 }