+++ service/controller/resource/monitoring/remotewriteingressauth/doc.go
@@ -1,0 +1,4 @@
+package remotewriteingressauth
+
+// This packages creates the remote write secret
+// to be used to secure the prometheus remote write ingress endpoint.